refactor: optimize dish
