+++ apps/mobile/service/index.ts
@@ -1,4 +1,4 @@
-export { default as UserService } from './user-service';
+export { default as UserService } from './user.service';
 export { default as CategoryService } from './category-service';
 export { default as SSEService } from './sse.service';
 export { default as TagService } from './tag.service';
